refactor: вынести Slider в отдельный компонент
