+++ src/components/CardsSlider/slider/__button/_next/slider__button_next.css
@@ -1,0 +1,4 @@
+.slider__button_next {
+  margin: 0 auto 0 0;
+  transform: rotate(180deg);
+}
