+++ src/components/CardsSlider/slider/__button/slider__button.css
@@ -1,0 +1,17 @@
+.slider__button {
+  width: 49px;
+  height: 49px;
+  margin: 0 9px 0 auto;
+  background-color: rgba(0,0,0,0);
+  background-image: url('../../../../images/slider__button.svg');
+  background-position: center;
+  background-size: 41px;
+  background-repeat: no-repeat;
+  border: none;
+  cursor: pointer;
+  transition-duration: .7s;
+}
+
+.slider__button:hover {
+  opacity: .77;
+}
