+++ src/components/CardsSlider/slider/__buttons/slider__buttons.css
@@ -1,0 +1,4 @@
+.slider__buttons {
+  display: flex;
+  margin-bottom: 28px;
+}
