+++ src/components/CardsSlider/slider/__item/slider__item.css
@@ -1,0 +1,7 @@
+.slider__item {
+  margin: 0;
+  padding: 0;
+  transition-duration: .7s;
+  background-image: linear-gradient(rgba(0,0,0,0), rgba(150,150,150,1));
+  border-radius: 30px;
+}
