+++ src/components/CardsSlider/slider/slider.css
@@ -1,0 +1,8 @@
+.slider {
+  display: flex;
+  width: 100%;
+  margin: 32px 0 40px;
+  padding: 0 0 0 132px;
+  column-gap: 24px;
+  list-style-type: none;
+}
